[IMP] Refactor code
