+++ odoo/js/loaders/frontend.js
@@ -1,16 +1,16 @@
 // Copyright 2019-2020 Alexandre Díaz <dev@redneboa.es>
 // License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-odoo.define("terminal.FrontendLoader", function(require) {
+odoo.define("terminal.loaders.Frontend", function(require) {
     "use strict";
 
     require("web.dom_ready");
     const core = require("web.core");
-    const Terminal = require("terminal.Terminal").terminal;
+    const Terminal = require("terminal.Terminal");
 
     // Ensure load resources
-    require("terminal.CoreFunctions");
-    require("terminal.CommonFunctions");
+    require("terminal.functions.Core");
+    require("terminal.functions.Common");
 
     $(() => {
         // A generic try-catch to avoid stop scripts execution.
